scaffold a modular project structure
