+++ CMakeLists.txt
@@ -24,7 +24,7 @@ ELSE (LIBBCC_LIB)
   MESSAGE(SEND_ERROR "libbcc not found")
 ENDIF (LIBBCC_LIB)
 
-SET(SOURCE_FILES h2olog.cc)
+SET(SOURCE_FILES h2olog.cc quic.cc)
 
 ADD_EXECUTABLE(h2olog ${SOURCE_FILES})
 TARGET_LINK_LIBRARIES(h2olog ${LIBBCC_LIBRARIES})
